refactor: move background checks into command/ to fix barrel cycle
Move skill-version-check, upgradeNotification, and version-check
into cli/utils/command/ as siblings of Base44Command and render.
This eliminates the circular dependency through the barrel that
caused Bun 1.3.11 to fail resolving runTask.
